Extract class ColumnFilterer.
